+++ frontend/src/pages/upload-link/file-uploader.js
@@ -6,7 +6,7 @@ import Resumablejs from '@seafile/resumablejs';
 import MD5 from 'md5';
 import { resumableUploadFileBlockSize, resumableSimultaneousUploads, maxUploadFileSize, maxNumberOfFilesForFileupload } from '../../utils/constants';
 import { seafileAPI } from '../../utils/seafile-api';
-import { clearFileUploadRuntimeState, consumeSuppressedUploadErrorToast, getBaselineSimultaneousUploads, markUploadConflictAutoRetry, maybeMarkFileFinalizing, maybeStartPendingUploadDuringFinalize, moveUploadToRetryState, resetUploadConflictAutoRetry, resolveUploadSuccessResult, restoreUploadConcurrencyIfIdle, shouldAutoRetryUploadConflict, trackUploadResponseStatus } from '../../utils/upload-finalization';
+import { clearFileUploadRuntimeState, consumeSuppressedUploadErrorToast, getBaselineSimultaneousUploads, getInitialSimultaneousUploads, initializeAdaptiveUploadConcurrency, markUploadConflictAutoRetry, maybeMarkFileFinalizing, maybeStartPendingUploadDuringFinalize, moveUploadToRetryState, noteAdaptiveUploadFailure, noteAdaptiveUploadRetry, resetAdaptiveUploadConcurrency, resetUploadConflictAutoRetry, resolveUploadSuccessResult, restoreUploadConcurrencyIfIdle, shouldAutoRetryUploadConflict, trackUploadResponseStatus, updateAdaptiveUploadConcurrency } from '../../utils/upload-finalization';
 import { Utils } from '../../utils/utils';
 import { gettext } from '../../utils/constants';
 import UploadProgressDialog from './upload-progress-dialog';
@@ -54,12 +54,14 @@ class FileUploader extends React.Component {
     this.loaded = 0;
     this.bitrateInterval = 500; // Interval in milliseconds to calculate the bitrate
     this.isUploadLinkLoaded = false;
+    this.adaptiveUploadCleanup = null;
 
     window.onbeforeunload = this.onbeforeunload;
   }
 
   componentDidMount() {
-    const simultaneousUploads = getBaselineSimultaneousUploads(this.props.simultaneousUploads || resumableSimultaneousUploads);
+    const configuredSimultaneousUploads = getBaselineSimultaneousUploads(this.props.simultaneousUploads || resumableSimultaneousUploads);
+    const simultaneousUploads = getInitialSimultaneousUploads(configuredSimultaneousUploads);
     this.resumable = new Resumablejs({
       target: '',
       query: this.setQuery || {},
@@ -86,12 +88,16 @@ class FileUploader extends React.Component {
 
     this.bindCallbackHandler();
     this.bindEventHandler();
+    this.adaptiveUploadCleanup = initializeAdaptiveUploadConcurrency(this.resumable, configuredSimultaneousUploads);
   }
 
   componentWillUnmount = () => {
     window.onbeforeunload = null;
     if (this.props.dragAndDrop === true) {
       this.resumable.disableDropOnDocument();
+    }
+    if (typeof this.adaptiveUploadCleanup === 'function') {
+      this.adaptiveUploadCleanup();
     }
   };
 
@@ -130,6 +136,7 @@ class FileUploader extends React.Component {
     this.resumable.on('fileProgress', this.onFileProgress.bind(this));
     this.resumable.on('fileSuccess', this.onFileUploadSuccess.bind(this));
     this.resumable.on('fileError', this.onFileError.bind(this));
+    this.resumable.on('fileRetry', this.onFileRetry.bind(this));
     this.resumable.on('uploadStart', this.onUploadStart.bind(this));
     this.resumable.on('progress', this.onProgress.bind(this));
     this.resumable.on('complete', this.onComplete.bind(this));
@@ -233,6 +240,7 @@ class FileUploader extends React.Component {
   onFileProgress = (resumableFile) => {
     const simultaneousUploads = getBaselineSimultaneousUploads(this.props.simultaneousUploads || resumableSimultaneousUploads);
     let uploadBitrate = this.getBitrate();
+    updateAdaptiveUploadConcurrency(this.resumable, resumableFile, uploadBitrate);
     let uploadFileList = this.state.uploadFileList.map(item => {
       if (item.uniqueIdentifier === resumableFile.uniqueIdentifier) {
         if (uploadBitrate) {
@@ -419,6 +427,7 @@ class FileUploader extends React.Component {
       this.retryUploadWithFreshLink(resumableFile, { resetAutoRetry: false });
       return;
     }
+    noteAdaptiveUploadFailure(this.resumable, resumableFile, message);
 
     let error = '';
     if (!message) {
@@ -496,6 +505,7 @@ class FileUploader extends React.Component {
     this.notifiedFolders = [];
     // reset upload link loaded
     this.isUploadLinkLoaded = false;
+    resetAdaptiveUploadConcurrency(this.resumable, getBaselineSimultaneousUploads(this.props.simultaneousUploads || resumableSimultaneousUploads));
   };
 
   onError = (message, file) => {
@@ -576,6 +586,7 @@ class FileUploader extends React.Component {
   onCloseUploadDialog = () => {
     this.loaded = 0;
     this.resumable.files = [];
+    resetAdaptiveUploadConcurrency(this.resumable, getBaselineSimultaneousUploads(this.props.simultaneousUploads || resumableSimultaneousUploads));
     this.restoreConcurrencyIfIdle();
     // reset upload link loaded
     this.isUploadLinkLoaded = false;
@@ -621,6 +632,7 @@ class FileUploader extends React.Component {
       totalProgress: 100,
       uploadFileList: uploadFileList
     });
+    resetAdaptiveUploadConcurrency(this.resumable, getBaselineSimultaneousUploads(this.props.simultaneousUploads || resumableSimultaneousUploads));
     this.restoreConcurrencyIfIdle();
     // reset upload link loaded
     this.isUploadLinkLoaded = false;
@@ -628,6 +640,10 @@ class FileUploader extends React.Component {
 
   onUploadRetry = (resumableFile) => {
     this.retryUploadWithFreshLink(resumableFile);
+  };
+
+  onFileRetry = () => {
+    noteAdaptiveUploadRetry(this.resumable);
   };
 
   retryUploadFile = (resumableFile) => {